Do not assume negative age for record_eviction_age
- Revert back to the original behaviour of the `record_eviction_age`
function that assumes that the 'age' argument is non-negative.
We do not want to assume any incorrect behaviour for the case where
'age' is negative. If the simulator does arrive at the case where
'age' is negative, then the programmer running the simulator will arrive
at a seg fault. With the help of gdb, the programmer
running the simulator will then be reminded that 'age' should
be non-negative after looking at the code that seg faulted.